Added config overloads, case sensitive rename to correct naming
