+++ README.md
@@ -1,12 +1,12 @@
-Serilog.Sinks.ExceptionLess
+Serilog.Sinks.Exceptionless
 ===========================
 
-[![Build status](https://ci.appveyor.com/api/projects/status/bvmfe8muijhgkb9j?svg=true)](https://ci.appveyor.com/project/serilog/serilog-sinks-exceptionless)
+[![Build status](https://ci.appveyor.com/api/projects/status/bvmfe8muijhgkb9j?svg=true)](https://ci.appveyor.com/project/serilog/serilog-sinks-Exceptionless)
 
-Serilog sink to publish to ExceptionLess
+Serilog sink to publish to Exceptionless
 
 ```
-PM> Install-Package Serilog.Sinks.ExceptionLess
+PM> Install-Package Serilog.Sinks.Exceptionless
 ``` 
 
 Read the Exceptionless configuration
@@ -18,6 +18,6 @@ ExceptionlessClient.Default.Register();
 ```csharp
 Log.Logger = new LoggerConfiguration()
     .Enrich.With(new HttpRequestIdEnricher())
-    .WriteTo.ExceptionLess(b => b.AddTags("Cart").AddObject(order).AddRequestInfo())
+    .WriteTo.Exceptionless(b => b.AddTags("Cart").AddObject(order).AddRequestInfo())
     .CreateLogger();
 ```
