+++ Serilog.Sinks.Exceptionless.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assets", "Assets", "{D1B3AA3E-8543-4C53-86BA-2D9306C05812}"
 	ProjectSection(SolutionItems) = preProject
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.ExceptionLess", "src\Serilog.Sinks.ExceptionLess\Serilog.Sinks.ExceptionLess.csproj", "{1DC3D8DA-3743-411A-B7B4-2A4F11398A58}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.Exceptionless", "src\Serilog.Sinks.ExceptionLess\Serilog.Sinks.Exceptionless.csproj", "{1DC3D8DA-3743-411A-B7B4-2A4F11398A58}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
