+++ src/Serilog.Sinks.Exceptionless/Sinks/Exceptionless/ExceptionlessClientExtensions.cs
@@ -3,7 +3,7 @@ using Exceptionless;
 using Serilog.Core;
 using Serilog.Events;
 
-namespace Serilog.Sinks.ExceptionLess {
+namespace Serilog.Sinks.Exceptionless {
     public static class ExceptionlessClientExtensions {
         public static EventBuilder CreateFromLogEvent(this ExceptionlessClient client, LogEvent log) {
             var builder = log.Exception != null
