+++ src/Serilog.Sinks.Exceptionless/packages.config
